+++ src/app/[admin]/organizations/[organizationId]/courses/[courseId]/(courseTabs)/students/components/StudentDetailsView.tsx
@@ -7,8 +7,8 @@ import { Badge } from '@/components/ui/badge'
 import { Input } from '@/components/ui/input'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import { ArrowLeft, Mail } from 'lucide-react'
-import { useGetStudentInfo } from '@/hooks/useGetStudentInfo'
-import { useGetStudentCompletedClasses } from '@/hooks/useGetStudentCompletedClasses'
+import { useGetStudentInfo } from '@/app/[admin]/hooks/useGetStudentInfo'
+import { useGetStudentCompletedClasses } from '@/app/[admin]/hooks/useGetStudentCompletedClasses'
 import { toast } from '@/components/ui/use-toast'
 import { DataTable } from '../../../../../../../../_components/datatable/data-table'
 import { createAttendanceColumns } from './attendanceColumns'
