+++ src/app/[admin]/organizations/[organizationId]/courses/[courseId]/(courseTabs)/students/page.tsx
@@ -21,7 +21,7 @@ import AddStudentsModal from '../../_components/addStudentsmodal'
 import { columns } from './columns'
 import { useStudentData } from './components/useStudentData'
 import { ComboboxStudent } from './components/comboboxStudentDataTable'
-import { useGetCourseStudents } from '@/hooks/useGetCourseStudents'
+import { useGetCourseStudents } from '@/app/[admin]/hooks/useGetCourseStudents'
 import AlertDialogDemo from './components/deleteModalNew'
 import { DataTablePagination } from '@/app/_components/datatable/data-table-pagination'
 import axios from 'axios'
